move periodindex comparison tests
